feat(monatsplaner): adapt operation to update Elterngeldbezüge
This makes it a more standalone operation on the same level as other
"entry-level" operations. In result it can be used also straight forward
by anyone externally and no more just within the package itself.
This also had the advantage that some duplicated code that was necessary
to actually call this operation could be shifted into this operation
itself.

Furthermore, the interface and feature set of the operation has been
extended. It still allows to specify an Elternteil to make the updates
for. This is often the case as changes to the Plan happen usually per
Elternteil. But it has no the ability to also just update it for
all Elternteile together. This is computational heavy and wanna be
avoided if not necessary.