refactor(planer): extract Anleitung into own component
The Planer component is a bigish entry component that assembles all the
top-level component that make up the Planer and connects them with the
service. The whole visual stuff like text, lists etc. of the initial
instructions clutter the component and hurt its readability. Therefore,
we follow our strategy to encapsulate such low-level visual components
to keep it clean.

This also allowed to add some simple tests for this part. Though, they
are very trivial to be honest. Anyhow, in result a small bug was found
due to a weak type usage. Happily it was no production bug.