feat(planer): allow to deselect an Option again
This is actually an anti-pattern. Theoretically a radio button group
must always have one radio checked. Having it initially empty is an
exception. For later unchecking the "kein Elterngeld" Option is given.
However, it was decided that this will be a big boost in user experience
if it is possible to uncheck an already checked radio again to reset the
whole group. This is not natively supported and only works with
controlled HTML input elements with some "trick".

From a domain perspective, this is actually not choosing an Option. It
is rather a new use-case/operation. However, because the chosen Option
will change, this requires to trigger also all the automation
implemented for choosing an option (i.e. Partnerschaftsbonus). This
automation is currently not re-usable and tightly coupled to the
respective operation functions. There is also no straight solution to do
so. In that context, the chose Option operation was extended to also
serve this new use-case scenario. In practice it was just a type thing.

Some test cases had to be adapted to respect the automated
Partnerschafsbonus selection. As it is no more safe to click and already
selected option, it must be prevented. Tests failed because
Partnerschaftsbonus was selected, which made the following Lebensmonat
being automatically selected. Clicking no the Partnerschaftsbonus in the
following Lebensmonat will reverse the complete selection, also in the
previous Lebensmonat.