feat(planer): improve Beispiele to merge better with Mutterschutz
The old logic was just overwriting whole Lebensmonate based on the
initial Lebensmonate of the domain. This especially negated any choices
for the partner of the mother. The new solution allows to properly merge
whole Lebensmonate.

This fixes an issue with one specific Beispiel we have defined. An
additional "typo" has been fixed too.