feat(monatsplaner): "improve" finde letzten Lebensmonat operation
The former operation had an `OrDefaut` postfix. This is not necessary.
Enhancing the return type to express the possibility for an
potentially "negative" case is good enough. It is up to the caller to
decide what should happen in that case.

Furthermore, the implemented logic was not fully correct from a domain
perspective. Moreover, it excluded the possibility of Monate with only
Bruttoeinkommen.

It is a little unclear if that should be actually within the
`monatsplaner` package at all.