feat(planer): fully show Beispiel Pläne
The Planer does initially show just the first 14 Lebensmonate of the
Plan. The user is then able to show more if needed (and reverse). With
the feature extension of the Beispiele, there is a new requirement. When
selecting a Beispiel, all planned Lebensmonate of the Plan must be
visible. When switching between Beispiele it must continuously adapt to
the number of planned Lebensmonate. This is important to properly
communicate the Beispiele to the user.

In result, the whole expand-collapse logic got more complex, as it not
only depends on a single component internal state, but also the latest
Plan passed as property. The latter is affected by continuous changes.
Either caused by user interactions to the Plan itself or other factors
like the Beispiele.
In that regards, it was much simpler to re-implement this logic from
scratch using test-driven-development. This is mentioned to explain the
diff of the commit as it is. The desired behavior has been double
checked and refined with a designer again.