refactor(monatsplaner): move finde letzten verplanten Lebensmonate to planer feature
This operation is strictly speaking not part of the Monatsplaner domain.
To keep the domain package focused we wanna try to avoid including
special purpose logic for the user-interface. Therefore, we already
moved plenty of other code too. This is always a slightly tricky balance
to maintain.

In result, the Lebensmonateliste component moved into a directly to
include the module for the moved operation function. In turn, the whole
LebensmonatDetails component tree moved too. We wanna try to keep the
component tree within the directory structure of the module to remain
organized an have proper scopes. In that context, notice that the
context for the Lebensmonat Information makes some trouble and might
see some refactoring too.