feat(planer): add extra hint to explain disabled Auswahloptionen
We discovered that is isn't always 100% clear how the whole mechanism of
selectable and not selectable Auswahloptionen works. Therefore, an extra
hint pops up for Lebensmonate that have any not selectable
Auswahloption. The hint should help the user to understand how to gain
further information.

The current implementation is very inefficient. Yet we have no
performance issues and its a bit tricky to improve properly.