refactor(planer): maintain components directory structure
The Planer feature continues to grow, especially the components
directory. To continue maintaining a structure that is easy to follow
and communicative, some changes were necessary.

This commit focuses on some common components that do not fit into the
component tree. Though, they are special and for the feature only.
Leaving the in the root directory of the features components makes this
folder cluttered and harder to grasp. Therefore, we wanna follow a new
pattern to put very basic components that are shared cross a whole
feature within a common directory.

Furthermore, some minor import tweaks have been added to keep each
component modules imports short and reasonable. There is still one
exception of the context module used for the whole LebensmonateDetails
sub-tree.