refactor(planer): maintain entry components
This is a follow up to clean up the heavily grown Planer feature
components directory. It should be easier to follow and use. Therefore
the cluttered root directory of the component for this feature need to
be better structured again. Thanks to the almost clean component tree
structure, the movement of all the modules if very straight forward.

The feature started to develop some "sub-features" with multiple entry
points of components that it exposes. This must be more obvious from the
code structure. Having this sorted out, its now much easier to grasp
again what belongs together. Its less stressful to work with and makes
the decision where to place something more straight forward.