refactor: shift common components closer to clients
After the introduction of a strong feature based application structure,
some component modules remained in the application shared component
directory. It was mentioned that this will need to see some clean up.
The goal is to properly place each component where it belongs. It should
be avoided to make a component widely shared if that is not necessary.
There will be component that represent a very basic concept and are
designed to be highly re-used. But such component, or any such code,
must be designed quite generically and fostered to its core concept. It
must be avoided to have it being serving too many clients that have
different requirements. A good example where this went wrong is the
CurstomRadioGroup (which will need some refactoring) or even the Button.

This commit is just about iterating over all application shared
components and identify their optimal location. Some were placed very
locally to their single usage in the component tree of a feature. Other
have moved to become a commonly shared component within a feature. And
the rest remains as truly shared components cross all features.

In future, components might move up the directory tree. If it fits and
is applicable, they can become common components of their feature or
application wide. This is a natural process. It potentially will require
some sharpening of the respective component during this journey.

The remaining application wide common components are not in a good
shape. They will need some refactoring too. Else their ability to be
adopted widely is actually limited.