refactor: organize directory naming for module interfaces
The projects source directory currently consists of 388 files. Including
tests as in-source and spec files. This does not include other packages
like for the metrics synchronization. Just to mention the scope.
Especially in such a project that suffers further modularization, it is
important that each aspect of the application is modelled with a clean
interface in mind. So it should be avoided to use deep import locators
where possible and use module index files that provide a clear interface
for its respective source sub-tree. For example the application features
clearly expose what they wanna expose. The client of the feature must
not index deeply into it.
Notice that some mocking mechanisms are still required in absence of
other dependency injection mechanisms. Such spies/mocks need to be focus
on single modules to limit side-effects.

In the context of the above matter, a pattern of directory naming has
been introduced. This is best explained by an example case. Having
(React) web-components, they often start small and simple. After
some iterations they became so big, it is appropriate to split them into
multiple source files. To maintain a clear interface, the component uses
a new directory to live in with all its custom code. Using an index
module allow it to expose only itself to the outside.
However, component files use pascal case naming, just as the render
function itself. Source directories usually don't. So it casually
happened that the name was transformed into kebab case. In result, all
clients referring to it had to update their locators. This should be
avoided. "Transforming" a module in such a manner back and fourth should
have no impact on them.

Furthermore, such naming is more communicative. Naming directories like
their primary code unit resident helps to communicate scopes. If
a sub-directory of components for a feature for example uses the typical
naming for a web component, it makes very obvious that nothing inside is
of interest of someone external. Just what it decides to expose should
be used. Everything else must have no stable interface as it is
considered local and volatile.

The latter has a special meaning when applied to the domain model of the
Monatsplaner. The domain model is designed in a functional style.
Thereby there are no methods for use-cases of entities and object
values. But they are separated operations and specifications. Anyhow, to
the outside everything is scoped under the model name they belong to.
The internals should be not transparent.

Unfortunately there are no good linting rules or architecture fitness
tests suitable for this. The concepts are not easily expressed
statically and currently require decision making by engineers and
careful coding. Anyhow, at least the import locator usage might be
supported by special linting soon.