style(application): unify definition of web component properties
This is purely a measurement to unify the code style in the code base.
A unified style helps the reader to grasp the details of the actual
implementation and is not strained by continuously see the same pattern
in different styles. That makes it harder to see the important stuff.

First of all, the definition of the properties type can be a plain type.
There is no need to use interfaces in this context. It is more heavy on
the compiler side and usually provides less usable editor hints as used
here. After all its more clean that the same approach is used
everywhere.
Anyhow, it is absolutely possible that in some future a component needs
to define its properties as an interface. If that happens it will nicely
stand out as it is an "exception" and is thereby better to catch while
reading.

The type definition of the properties is basically always local to the
module of the render function. Therefore, it is not necessary to scope
the name for the component. It just makes it harder to read, especially
as the "Props" keyword comes at the end.