fix(planer): fix reactivity rendering of printing view
There was the desire to provide a unified printing experience. Therefore
we use a hidden component that gets rendered. Though, for printing it is
not hidden, but everything else. The issue was that the printing view
was lazy. Should mean on printing, it did not show the latest/current
Plan.
The Plan provided to the printing component was simply not reactive. It
was based on a reference which does not trigger rendering on mutation.
Based on past requirements a reference was more appropriate and
efficient. A quick fix is to replace this with a reactive state instead.
However, it might be worth to investigate if a more explicit mechanism
might be more clean. Its currently "hard" to grasp from the code how this
mechanism actually works as it is very implicit.