feat(planer): add tracking for selection of Beispiele
This commit has two significant changes in it.
First of all, the identifiers for the Beispiele changed. In the past
they were dynamically generated. The solely purpose was to differentiate
the Beispiele in a set of generated sets for state management. However,
for the tracking it is interesting to know which Beispiele are chosen.
The remaining describing properties are not unique enough as Beispiele
for different user groups use the same/similar descriptions. Therefore,
the identifier for each individual Beispiel is now static. So it can be
used for tracking. For simplicity and readability without an extra
parsing steps, the identifiers are plain text based on their context.
This allows to communicate user group and exact Example without any
further knowledge or mechanism.

Furthermore, the whole callback collection for the Planer feature have
been extended to cover the Beispiele service too. Though, yet on
a smaller scale, the same design as for the planer service has been
chosen. In first place to have a uniform pattern. However, the splitting
in the external interface of the Planer component starts to get a bit
messy. It's hard to properly separate all the merged callbacks. However,
the properties should remain somewhat simple.

The related data layer variable, event trigger and tag have been created
in the Matomo tag manager.