feat(planer): improve tracking around usage of planer
This adds a new tracking event that only fires when the user is changing
the Plan by selecting an option. This provides some more focused
tracking.
Additionally, this can be combined with the tracking variable of the
identifier for a selected Beispiel. It allows to derive more knowledge
how the users combine these two.