chore(planer): extract tracking from planer page component
The planer page is responsible to plug everything together. This
includes to inject tracking for certain events of the planer, which
inverses this responsibility.
However, the components render function got really cluttered and hard to
maintain/read. Therefore, all tracking related stuff was exported into
a module within the components directory.