test: replace module mocking with spies
The mocking of whole module has led to multiple issues in the past.
Especially in the context of the still young support of in-source tests.
However, using spies is much less invasive. It has multiple advantages.
One is, that its not necessary to mock deeply into the directory tree.
This allows to respect the boundaries of the code base. So far, the
mocking imports were some kind of exception to the import rules.
Moreover, it is possible to just mock and alternate behaviors when
necessary. This allows to selectively apply mocking for single tests.
In general its much more controlled with less side effects to worry
about. It even allowed to enable some skipped tests that did not work in
the past.