refactor(application): simplify button as styled component
The Button component is to 99.9% just a styled component. However it
hardly tries to replicate the interface of a regular button element. To
simplify this and make it much more "natural", this component literally
just became a styled component.
There are considerations to get rid of such kind of components in favor
of CSS "components".