feat(user-tracking): extends Nutzergruppen for werdende Eltern
The huge majority (>80%) of all users are categorized as "werdende
Eltern. To gain further insides, this group was split into two.

The implementation of the mapping function has been replace because it
was not well suited anymore. It was my own implementation, but it was
not too good. It also uses the date utility functions we use everywhere
else.
The testing has been changed to a property based approach for better
guarantees. It was necessary to provide some convenience interface for
the data generation to improve the readability and expressiveness of the
tests.