feat(abfrageteil): final clean up of remaining info dialogs
It was decided that we don't wanna use dialogs anymore for the
Abfrageteil. At least not to provide additional information to
understand the forms. Instead we use details elements (called info
texts...).
The majority of infos has been extracted into purely visual components.
This should help to keep the form components more tidy and terse.

The infos were formerly heavily injected into awkward custom input
elements. The current approach is to use more generic slots.
There was a lot of clean up work everywhere. The spacing is a little
chaotic and was tried to be unified better.
Finally, the InfoDialog component has moved into the planer feature, as
it is solely used here now.