refactor(application/abfrageteil): simplify date input
Following standards is usually a good idea. The react-hook-form
dependency claims to embrace native forms and inputs. Also custom input
component should embrace these. In result, they can be (re-)used very
easily and interchangeably. And they'll be independent of
react-hook-form, but even integrate better with it right now.

The current custom input components take a lot of responsibility and
have a lot of dependencies. This is a first commit to start to purify
the CustomDate component. The component still remains is responsibility
to show error/validation related state. This is a big pattern in the
Abfrageteil that will need more attention independently.