refactor(application/abfrageteil): remove custom input component
This is another commit in the context to establish standard HTML input
element interfaces for simplicity and interoperability.
For this input, it is actually just some styling of an input element.
Nothing a custom component should be used. At the moment, an absolute
trivial approach to share the styling has been chosen. We might soon
wanna shim into the idea of using some style dictionary or similar
solution for shared styling for certain elements.