feat(elterngeldrechner): introduce history for kassenindividuellen Zusatzbeitrag
Like already the Werbekostenpauschale, the kasseninvidueller
Zusatzbeitrag changes over time. The value to use for the calculation
depends on the birthdate of the child. Therefore, the beforehand
extracted logic to represent the history of parameters has been used.
However, so far only a single historic entry exists, containing the
former constant value. Next step is to fill the history.

The derivation/origin of the historic value is actually quite
"interesting" as it conducted by a chain of laws and norms. The related
knowledge is therefore documented within the module.

Furthermore, there is some "translation" going on. The
Programmablaufpläne do expect a specific format of the parameter.
However, it should not dictate the format to the Elterngeldrechner. So
a mathematically better (?) representation was chosen for the history.
For the calculation with the Programmablaufplan the value is
arithmetically translated, using the already existing precise operation.