feat(elterngeldrechner): fill history for kassenindividuellen Zusatzbeitrag
In result, we get some issues with the tests again. The former
"constant" value was for the year 2015. In result, some tests were
simply no more correct for their respective birthdate. Because the
support of Lohnsteuerjahre is limited, the values has to be adapted
optimistically according to the latest calculation.

Also the original-rechner required an exception to still make any sense
to be used for the refactoring tests. However, the possible birthdates
must be reduced even further due to this.