refactor(elterngeldrechner): remove geschwisterbonus from many results
The Geschwisterbonus gets calculated in multiple locations and is
forwarded via multiple calculation result data structure. The actual
important part is when the Elterngeldausgabe for a Lebensmonat is
calculated. While this could be erased too, this is the only final
instance remaining where it gets communicated in the output if the
Elterngeld contains the bonus or not.

The goal is to reduce complexity and error sources. There are in general
too many values that get calculated multiple times on different stages.