feat(elterngeldrechner): adopt Geburtstag for persönliche Daten
Further integration of the new domain model. This should help to avoid
date comparison issues, especially in property tests.

Notice that also the naming of the property has changed. From the
perspective of the Elterngeldrechner domain, the birthday is not
"probably expected". It is just the birthday of the child the Elterngeld
is calculated for. This differs from the Abfrageteil feature of the
application, which has a different context.