feat(elterngeldrechner): wrap up Geburtstag usage
This finishes the introduction of the new Geburtstag domain model. In
result, also a consequent naming of properties, parameters etc. is used
throughout the Elterngeldrechner now.
There are a very few last instances of the native Date type in use.
These are on purpose (for now?).

Notice that other packages/domains/features are still untouched.