feat(elterngeldrechner): add improved implementation of Geschwisterbonus
This is part of the big "testing" strategy (level 1). It's a more
thoughtful implementation of the basic building block for the
Geschwisterbonus. With a different approach, the code is almost
identical to the law text. Furthermore, great effort has been made to
test this really throughout.
However, due to current limitations of the higher level calculation
flow, the bonus amount calculation is incomplete for now.

The complexity for the tests is quite big. There are a lot of tweaks
that try to provide readable and terse test cases that can be followed.
Therefore multiple abstraction levels are introduced to help. There is
definitely some optimization left, but hasn't been implemented to avoid
too much special abstractions.