refactor(elterngeldrechner): exchange Geschwisterbonus logic
Remove the old part and inject the just added re-implementation.
Due to the different solution approach to determine if Geschwisterbonus payed
out for a Lebensmonat or not, some result properties had to be deleted as the
data is simply not there anymore. In result some related code code be clean up
and simplified a little.
Notice however, that due to the broken higher level calculation flow, the
Geschwisterbonus can't be calculated for any Lebensmonat with ElterngeldPlus
yet.