feat(application/planer): further integrate with automatic Bruttoeinkommen
With the recent feature addition, when choosing Partnerschaftsbonus it
will automatically check if there is already another Monat with
Partnerschaftsbonus that has some Bruttoeinkommen defined and applies it
automatically. This should be extended to also trigger when some
Bruttoeinkommen is inserted for the first Partnerschaftsbonus Monat to
automatically fill it to all other Monate with Partnerschaftbonus that
have no Bruttoeinkommen yet. This is especially useful, because there
will be always a second Monat with Partnerschafsbonus selected
automatically.
However, this intergration is a little tricky. The most minimal approach
would be to trigger this operation when setting some Bruttoeinkommen.
That would leave this logic fully enclosed in the `monatsplaner`
package. Unfortunately, this does not work out. The user interface
triggers the operation to set a Bruttoeinkommen (basically) on every
keystroke of the user. That means once the first number is typed, the
automation fills it as Bruttoeinkommen to other Monate with
Partnerschaftbonus. On continuous typing, this will do nothing anymore,
because the other Monate already have some Bruttoeinkommen now.
A solution to fix this needs some kind of "input is finished". The most
straight forward approach would be to trigger the operation of the input
on leaving the input. However, this makes the user interface much less
reactive and users can't interactively see how their Elterngeld adapts
automatically to the input. In result, it was decided to decouple these
two operations. On every keystroke, the Bruttoeinkommen is updated with
a responsive planer. Once the users leave the input, the automatic
operation to fill other Monate with Partnerschaftsbonus is triggered.
To implement this feature, it can't be implemented in the `monatsplaner`
alone, but requires cooperation with the user interface of the
application. In result, a lot of code has been added to provide this
operation to the user interface down to the input for the
Bruttoeinkommen. It requires ugly prop drilling and way too much code.
It's really not a nice solution. It's another score into the technical
debt to find an improved solution approach for this. Anyhow, the feature
works well and the user experience should be highly improved.