chore(monatsplaner): clean up obsolete reset operation for Monat
Some while ago, we added the feature that the operation to choose an
Option for a Monat allows to "reset" it again. Due to semantic reasons,
this wasn't the case before. Therefore, an extra operation was provided
to do a reset. This is no more necessary and could be replace with calls
to the choose Option operation.

This includes some additional renaming as `setzeOptionZurueck` in
singular was not working too well anymore. Actually it never did.
Hopefully, with the new naming the operations purpose is a little more
clear.