feat(monatsplaner): include Bruttoeinkommen when resetting a Monat
When resetting a Monat to have no Option chosen the Bruttoeinkommen gets
reset too. Because in practice it usually makes a lot of sense and is
more intuitive. If the user decides to take "kein Elterngeld" the
Bruttoeinkommen remains. But deselecting any option is like reset
everything for this Monat. This is especially the case for more complex
automated scenarios like with Partnerschaftsbonus. Having multiple
Lebensmonate with Partnerschaftbonus and respective Bruttoeinkommen
become pretty weird once the user decides to choose a different Option
and in result all/many Monate with Partnerschaftsbonus are reset. Having
their Bruttoeinkommen being cleared too is very useful.