feat(elterngeldrechner): documented rewrite of Abzüge for Sozialabgaben
The next big chapter in revising the Elterngeldrechner are the
Sozialabgaben. Though, it is really big. So this is just the very first
tiny step.
The new module has been put into new directory to take into account the
grow of the related logic all around the Sozialabgaben. The new
functions aren't anything special or are more correct. Though, the are
explicit functions that are well named, reference important sources and
have some dedicated property tests.