feat(elterngeldrechner): add revised implementation for Midi-Job
This is another brick in improving in context of the Sozialabgaben. The
calculation for the part of the income that is relevant for these
deductions was integrated into the higher level algorithm. Furthermore
there were two occurrences with different implementations.
Following the new approach, the revised implementation uses terminology
of the regulatory norms, has a more aligned implementation with
original, related documentation and dedicated tests.