feat(elterngeldrechner): integrate with new Midi-Job calculations
This simply makes use of the revised calculation units of the algorithm
of the previous commit.

The higher level algorithm can't yet make use of the historic aware
calculation. Therefore it would need to know the accurate
Bemessungszeitraum and income per month to follow the law correctly.
However, a intermediate solution has been established that is properly
documented in code.

There were some serious testing issues again. The original-rechner did
not only use obsolete parameter values, but also a combination of values
that was never valid at the same point in time. It was necessary to fix
these parameters. Though, the implementation itself wasn't touched.
Furthermore, the FIT tests continue to slip more and more. More cases
get ignored as I couldn't find a balance to address all the inaccuracies
at the same time.