chore(elterngeldrechner): follow more strict file name casing
This is a concept we started to use in the code base to name things by
the domain properly. And as German nouns are written upper case, the
file names do too. In result it's much easier to read and there is no
special distinguishment made.

It has the convenient side effect that already revised and "old" modules
are separated visually. As many file-system viewers sort upper case names
before lower case. Though, this must not be case and is not the primary
intend at all. However, it can be a tiny bit helpful.