chore(elterngeldrechner): improve correct presentation of Steuerklassen
This commit helps to bring the code closer to the domain.

First of all, the variants of the enumeration have changed in their
naming and values. They now align fully with regulatory texts of all
relevant norms. It also reads a little better as the name of the
enumeration is not repeated in the names of the variants.

Some minor "documentation", especially for the special variant of
Steuerklasse IV with a factor.

Dissolve encode and decode functions from the domain. This functionality
is used and dependent in context of external dependencies. Like the
encoding used in the tests sheet by the FIT, but also the format for the
Programmablaufpläne by the BMF.

Consistent naming everywhere without artificial casing. I suspect this
"technique" was used to improve the readability of longish German nouns.
However, we have moved away from this as it rather hurts the readability
from a domain perspective.