chore: refactor tsconfig to explicitly set outDir
As a safety feature VS Code prevents the compiler from writing accidentally into the src/ folder
while constantly running type checking in the background.
In order for the background checks and tsc to behave correctly "outDir" is set to "dist"
and also added to the excluded directories together with "node_modules".