docs: add domain section to root readme.md
The elterngeldrechner requires engineers to have a basic
understanding of the relevant legal requirements in order
to make changes to its core functionality. This section is
the result of a knowledge exchange meeting with a former
engineer and represents an effort to better integrate
domain knowledge with the code. I believe that information
relevant to engineer is kept close to code, if not even
as code.