refactor(planungshilfen): inverse beispiel type definition
Exposing the type as BeschreibungFuerEinBeispiel does contain
implicit knowledge about how the feature is beeing used and
also clashes with the contained property beschreibung.

I've inverted the type definition so that we know export a
version of the Beispiel which is agnostic to the Ausgangslage
and does not require knowledge of the type from the consumer.