refactor(planungshilfen): lift up useBerechneElterngeldbezuege
In order to provide the Gesammtsumme of a Beispiel the feature needs
access to the elterngeldrechner. Since the planer page has already
implemented a hook to access the elterngeldrechner the easiest way was
to lift it up to @/application/hooks.

Along the way I discovered two .spec.ts files dangling around which
would usually be in-source tests and refactored them.