refactor: group pages by abfrage and planung
During the implementation of the beispiele or planungshilfen I faced the
challenge of re-using the useBerechneElterngeldBezuege hook. After
moving it to the global hooks directory initially I recognized that it
depends on the global state beeing filled within the abfrageteil and
thus should not be shared globally. Though it can be shared by all pages
after the abfrageteil which was my motivation for this grouping.