feat(planungshilfen): implement calculation of elterngeldsumme
The calculation of the gesamtsumme was added to the
useBerechneElterngeldbezuege hook because it is mainly a composition of
the existing functionality for each elternteile.

The next step is to reduce duplicate logic between the header of the
planer and the beispiel page in order to prevent differences in the
gesamtsumme between the beispiel page and the planer header.