feat: improve link texts in datenuebernahme
After rolling out the Verzahnung, the Datenübernahme in den Antrag,
Feature, we realized that the click count is higher for the link to the
Anlagen than to the actual Antrag. The first hypothesis was that this
stems from a wording issue where user believe that the second link is
the correct one.