refactor(planungshilfen): remove beispiel service from page
The beispielService was primarily responsible for managing the active
state and providing a list of examples without the attached plan. State
management is now handled directly in the page for simplicity. The list
of examples without a plan was removed, as the new feature requires the
plan to be present.

Dissolving the useBeispielService hook comes with the downside that most
of the hook logic that is used by the page now lives in the feature
component directory meaning that the cohesion of the code is reduced.