refactor(planungshilfen): berechneElterngeldbezuege in erstelleBeispiele
The original implementation of gesammtsumme for planungshilfen was added
in the beispiel page as part of the user interface. To achieve this I
added a new method to the useBerechneElterngeldbezuege hook.

The addition to the hook was not optimal, as I found out in a call with
the former engineer. We started looking for another solution when we saw
that the Elterngeldbezug.ts file now contained another type,
BerechneElterngeldbezuegeByPlanCallback, which was only used within the
beispiel feature.

I learned that BerechneElterngeldbezuegeCallback was introduced to
invert dependencies and make the monatsplaner independent of the logic
for the calculation of elterngeldbezuege. I accidentally broke that
inversion by adding a type that was only needed for beispiele and
unrelated to the monatsplaner.

After discussing, we concluded that the field for elterngeldbezuege
within the Plan of a Beispiel should already be filled with the correct
value. This eliminated the need for planungshilfen to calculate it. With
this in place I reverted my changes to the useBerechneElterngeldbezuege
hook.

The beispiel feature now takes the berechneElterngeldbezuege function as
a parameter and applies it to each created plan.