feat(planungshilfen): grey span background on checked state
Implemented a grey background state for the relevant <span> elements
when the radio input is checked using tailwinds peer feature combined
with the [&_span] selector.

Due to tailwinds limitation peer-checked can only target sibling
elements. It was necessary to apply the style via a span selector inside
a wrapper. This may select more elements than intended in the future.

Refactoring is recommended if additional nested spans are added to avoid
unintended styling.